third_party/WebKit/LayoutTests/http/tests/security/mixedContent/nonwebby-scheme-in-iframe-allowed.https.html

<!DOCTYPE html>
<html>
<body>
<p>
    This tests that non-webby URLs are not blocked as mixed content. The test passes if a mixed
    content warning is displayed and the load proceeds. The test fails if a mixed content error
    is displayed, and the load is blocked.
</p>
<script>
if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.dumpChildFrames();
    testRunner.overridePreference("WebKitAllowRunningInsecureContent", false);
}
</script>
<iframe src="nonwebbyscheme://this-will-fail-but-no-mixed-content-error-should-appear"></iframe>
</body>
</html>
